not ended

+++ includes/libft.h
@@ -171,7 +171,7 @@ char					*ft_new_null_str(size_t size);
 char					ft_range_str(char *str);
 void					ft_swap_byte(void *x, void *y);
 void					ft_swap_n_bytes(void *a, void *b, size_t n);
-void					*ft_jump_pointer_to_n(void *pointer, size_t n);
+void					*ft_jump_pointer_to_n(void *pointer, int n);
 void					ft_sort_quick(void *arr, size_t size, size_t amount, int (*f)(void *, void *));
 char					*ft_str_disnuller_end(char *str);
